user_service/user_service/requirements.txt
alembic==1.8.1
aniso8601==9.0.1
autoenv==1.0.0
blinker==1.5
click==8.1.3
Flask==2.1.3
Flask-DotEnv==0.1.2
Flask-Login==0.6.1
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
passlib==1.7.4
psycopg2-binary==2.9.3
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.39
Werkzeug==2.1.2
zipp==3.8.1
